requirements.txt
Flask==3.0.3
plaid-python==24.0.0
python-dotenv==1.0.1
pandas==2.2.2
sqlite-utils==3.36
